+++ R/legend.R
@@ -97,10 +97,17 @@ addLegendImage <- function(map,
     htmlTag = list(htmlTag),
     f =
       function(img, label, htmlTag) {
-        fileExt <- tolower(sub('.+(\\.)([a-zA-Z]+)', '\\2', img))
-        stopifnot(fileExt %in% c('png', 'jpg', 'jpeg'))
-        htmlTag(
-          htmltools::tags$img(
+        if ( inherits(img, 'svgURI') ) {
+          imgTag <- htmltools::tags$img(
+            src = img,
+            style = 'vertical-align: middle; padding: 5px;',
+            height = height,
+            width = width
+          )
+        } else {
+          fileExt <- tolower(sub('.+(\\.)([a-zA-Z]+)', '\\2', img))
+          stopifnot(fileExt %in% c('png', 'jpg', 'jpeg'))
+          imgTag <- htmltools::tags$img(
             src = sprintf(
               'data:image/%s;base64,%s',
               fileExt,
@@ -109,9 +116,9 @@ addLegendImage <- function(map,
             style = 'vertical-align: middle; padding: 5px;',
             height = height,
             width = width
-          ),
-          htmltools::tags$span(label, style = labelStyle)
-        )
+          )
+        }
+        htmlTag(imgTag, htmltools::tags$span(label, style = labelStyle))
       }
   )
   if (!is.null(title)) {
@@ -164,58 +171,115 @@ addLegendImage <- function(map,
 #'
 makeSymbol <- function(shape, width, height, color, fillColor = color,
                        opacity = 1, fillOpacity = opacity, ...) {
+  strokewidth <- 0
+  if ( 'stroke-width' %in% names(list(...)) ) {
+    strokewidth <- list(...)[['stroke-width']]
+  }
   shapeTag <- switch(
     shape,
     'rect' = htmltools::tags$rect(
+      id = 'rect',
+      x = strokewidth,
+      y = strokewidth,
       height = height,
       width = width,
       stroke = color,
       fill = fillColor,
-      opacity = opacity,
+      'stroke-opacity' = opacity,
       'fill-opacity' = fillOpacity,
       ...
     ),
     'circle' = htmltools::tags$circle(
-      cx = height / 2,
-      cy = height / 2,
+      id = 'circle',
+      cx = height / 2 + strokewidth,
+      cy = height / 2 + strokewidth,
       r = height / 2,
       stroke = color,
       fill = fillColor,
-      opacity = opacity,
+      'stroke-opacity' = opacity,
       'fill-opacity' = fillOpacity,
       ...
     ),
     'triangle' = htmltools::tags$polygon(
-      points = sprintf('0,%d %d,%d %d,0', height, width, height, width / 2),
+      id = 'triangle',
+      points = sprintf('%s,%s %s,%s %s,%s',
+                       strokewidth,
+                       height + strokewidth,
+                       width + strokewidth,
+                       height + strokewidth,
+                       width / 2  + strokewidth,
+                       strokewidth),
       stroke = color,
       fill = fillColor,
-      opacity = opacity,
+      'stroke-opacity' = opacity,
+      'fill-opacity' = fillOpacity,
+      ...
+    ),
+    'plus' = htmltools::tags$polygon(
+      id = 'plus',
+      points = draw_plus(width = width, height = height, offset = strokewidth),
+      stroke = color,
+      fill = fillColor,
+      'stroke-opacity' = opacity,
+      'fill-opacity' = fillOpacity,
+      ...
+    ),
+    'cross' = htmltools::tags$polygon(
+      id = 'cross',
+      points = draw_cross(width = width, height = height, offset = strokewidth),
+      stroke = color,
+      fill = fillColor,
+      'stroke-opacity' = opacity,
+      'fill-opacity' = fillOpacity,
+      ...
+    ),
+    'diamond' = htmltools::tags$polygon(
+      id = 'diamond',
+      points = draw_diamond(width = width, height = height, offset = strokewidth),
+      stroke = color,
+      fill = fillColor,
+      'stroke-opacity' = opacity,
+      'fill-opacity' = fillOpacity,
+      ...
+    ),
+    'star' = htmltools::tags$path(
+      id = 'star',
+      d = sprintf('M %s z M %s z',
+                  draw_plus(width = width, height = height, offset = strokewidth),
+                  draw_cross(width = width, height = height, offset = strokewidth)),
+      stroke = color,
+      fill = fillColor,
+      'stroke-opacity' = opacity,
       'fill-opacity' = fillOpacity,
       ...
     ),
     'stadium' = htmltools::tags$rect(
+      id = 'stadium',
+      x = strokewidth,
+      y = strokewidth,
       height = height,
       width = width,
       rx = "25%",
       stroke = color,
       fill = fillColor,
-      opacity = opacity,
+      'stroke-opacity' = opacity,
       'fill-opacity' = fillOpacity,
       ...
     ),
     stop('Invalid shape argument.')
   )
-  utils::URLencode(
+  svgURI <-
     sprintf('data:image/svg+xml,%s',
-                           as.character(
-                             htmltools::tags$svg(
-                               xmlns = 'http://www.w3.org/2000/svg',
-                               version = '1.1',
-                               width = width,
-                               height = height,
-                               shapeTag
-                             )
-                           )))
+            utils::URLencode(as.character(
+              htmltools::tags$svg(
+                xmlns = "http://www.w3.org/2000/svg",
+                version = "1.1",
+                width = width + strokewidth * 2,
+                height = height + strokewidth * 2,
+                shapeTag
+              )
+            ), reserved = TRUE))
+  structure(svgURI, class = c(class(svgURI), 'svgURI'))
 }
 
 makeLegendSymbol <- function(label, labelStyle, ...) {
@@ -224,6 +288,25 @@ makeLegendSymbol <- function(label, labelStyle, ...) {
     htmltools::tags$img(src = shapeTag, style = "vertical-align: middle; padding: 1px;"),
     htmltools::tags$span(label, style = sprintf("vertical-align: middle; padding: 1px; %s", labelStyle))
   )
+}
+
+draw_plus <- function(width, height, offset = 0) {
+  x <- width * c(rep(c(.4, 0, .4, .6, 1, .6), each = 2), .4) + offset
+  y <- height * c(0, rep(c(.4, .6, 1, .6, .4, 0), each = 2)) + offset
+  paste(x, y, sep = ',', collapse = ' ')
+}
+
+draw_diamond <- function(width, height, offset = 0) {
+  x <- width * c( .5, 0, .5, 1, .5) + offset
+  y <- height * c(0, .5, 1, .5, 0) + offset
+  paste(x, y, sep = ',', collapse = ' ')
+}
+
+draw_cross <- function(width, height, offset = 0) {
+  a <- sqrt(2) / 10
+  x <- width *  c(a, 0, .5 - a, 0, a, .5, 1 - a, 1, .5 + a, 1, 1 - a, .5, a) + offset
+  y <- height * c(0, a, .5, 1 - a, 1, .5 + a, 1, 1 - a, .5, a, 0, .5 - a, 0) + offset
+  paste(x, y, sep = ',', collapse = ' ')
 }
 
 #' Add Customizable Color Legends to a 'leaflet' map widget
@@ -289,6 +372,14 @@ makeLegendSymbol <- function(label, labelStyle, ...) {
 #' @param decreasing
 #'
 #' order of numbers in the legend
+#'
+#' @param opacity
+#'
+#' opacity of the legend items
+#'
+#' @param fillOpacity
+#'
+#' fill opacity of the legend items
 #'
 #' @param ...
 #'
@@ -434,6 +525,7 @@ addLegendNumeric <- function(map,
                              tickLength = 4,
                              tickWidth = 1,
                              decreasing = FALSE,
+                             fillOpacity = 1,
                              ...) {
   stopifnot( attr(pal, 'colorType') == 'numeric' )
   rng <- range(values, na.rm = TRUE)
@@ -524,6 +616,7 @@ addLegendNumeric <- function(map,
                                        width = width,
                                        x = rectx,
                                        rectround,
+                                       'fill-opacity' = fillOpacity,
                                        fill = sprintf('url(#%s)', id)))
                            ),
                            Map(htmltools::tags$line,
@@ -561,11 +654,13 @@ addLegendQuantile <- function(map,
                               values,
                               title = NULL,
                               labelStyle = '',
-                              shape = c('rect', 'circle', 'triangle', 'stadium'),
+                              shape = c('rect', 'circle', 'triangle', 'plus', 'cross', 'diamond', 'star', 'stadium'),
                               orientation = c('vertical', 'horizontal'),
                               width = 24,
                               height = 24,
                               numberFormat = function(x) {prettyNum(x, big.mark = ',', scientific = FALSE, digits = 1)},
+                              opacity = 1,
+                              fillOpacity = opacity,
                               ...) {
   stopifnot( attr(pal, 'colorType') == 'quantile' )
   shape <- match.arg(shape)
@@ -592,7 +687,10 @@ addLegendQuantile <- function(map,
     color = colors,
     labelStyle = labelStyle,
     height = height,
-    width = width
+    width = width,
+    opacity = opacity,
+    fillOpacity = fillOpacity,
+    'stroke-width' = 1
   )
   orientation <- match.arg(orientation)
   if ( orientation == 'vertical' ) {
@@ -614,10 +712,12 @@ addLegendBin <- function(map,
                          values,
                          title = NULL,
                          labelStyle = '',
-                         shape = c('rect', 'circle', 'triangle', 'stadium'),
+                         shape = c('rect', 'circle', 'triangle', 'plus', 'cross', 'diamond', 'star', 'stadium'),
                          orientation = c('vertical', 'horizontal'),
                          width = 24,
                          height = 24,
+                         opacity = 1,
+                         fillOpacity = opacity,
                          ...) {
   stopifnot( attr(pal, 'colorType') == 'bin' )
   shape <- match.arg(shape)
@@ -630,7 +730,10 @@ addLegendBin <- function(map,
       color = colors,
       labelStyle = labelStyle,
       height = height,
-      width = width)
+      width = width,
+      opacity = opacity,
+      fillOpacity = fillOpacity,
+      'stroke-width' = 1)
   orientation <- match.arg(orientation)
   if ( orientation == 'vertical' ) {
     htmlElements <- lapply(htmlElements, htmltools::tagList, htmltools::tags$br())
@@ -651,10 +754,12 @@ addLegendFactor <- function(map,
                             values,
                             title = NULL,
                             labelStyle = '',
-                            shape = c('rect', 'circle', 'triangle', 'stadium'),
+                            shape = c('rect', 'circle', 'triangle', 'plus', 'cross', 'diamond', 'star', 'stadium'),
                             orientation = c('vertical', 'horizontal'),
                             width = 24,
                             height = 24,
+                            opacity = 1,
+                            fillOpacity = opacity,
                             ...) {
   stopifnot( attr(pal, 'colorType') == 'factor' )
   shape <- match.arg(shape)
@@ -666,7 +771,10 @@ addLegendFactor <- function(map,
       color = colors,
       labelStyle = labelStyle,
       height = height,
-      width = width)
+      width = width,
+      opacity = opacity,
+      fillOpacity = fillOpacity,
+      'stroke-width' = 1)
   orientation <- match.arg(orientation)
   if ( orientation == 'vertical' ) {
     htmlElements <- lapply(htmlElements, htmltools::tagList, htmltools::tags$br())