+++ NEWS.md
@@ -1,4 +1,13 @@
-# leaflegend 0.2.0
+# leaflegend 0.2.4
+
+* `addLegendImage` now supports using an svg URI from the output of `makeSymbol`.
+To supply a custom svg URI, add the 'svgURI' class to the character vector.
+
+* added more shapes to `makeSymbol`, `addLegendNumeric`, 
+`addLegendQuantile`, `addLegendFactor`, `addLegendBin`
+
+* Control the opacity of the legend shapes for `addLegendNumeric`, 
+`addLegendQuantile`, `addLegendFactor`, `addLegendBin`
 
 * `makeSymbol` now returns embeddable svg